Add QuestionLoader as source of filters
Uses a version of the ContentLoader class from the
admin app to get question data from the SSP
Content repo to populate the filters.

The transformation of this data into filter
view data is handled by the SearchFilters class.

The process of getting filter view data now goes
like this:

1. QuestionsLoader is used to access the G6
questions
2. static methods on the SearchFilters class are
used to transform this into filter data
3. the filter data is merged with data about state
for the view filter data generated on each request

Also includes the following changes:

- fixes for python 3 and better use of Requests
- make QuestionsLoader interface use keywords