Enforce URLs without trailing slashes
This URL works:
https://preview.development.digitalmarketplace.service.gov.uk/buyers-guide

This URL doesn't work:
https://preview.development.digitalmarketplace.service.gov.uk/buyers-guide/

We need to support both because the existing app uses URLs with trailing
slashes, which users may have bookmarked.

…redirects all URLs with trailing slashes to their equivalent,
non-trailing-slash brethren/sistren.

This means standardising on not having trailing slashes in our URLs,
which @TheDoubleK and I think is cleaner.

Fixes some of the issues found in [this story](https://www.pivotaltracker.com/story/show/94291320)