Finalize breadcrumbs (hopefully).
Talked to Ralph, who said we don't want breadcrumbs with the
page title on pages where the page title is explicit.
(The one exception being the 'Cloud technology and support' page
because that one's "kind of like a front page.")
Also have breadcrumbs when search bounded by lot so that it's
more obvious what's going on.