Test for presence of analytics in minified JS
Minifying the JS shortens variable names wherever their scope makes it possible.
This means that `GOVUK.analytics.trackPageView` gets shortened to something like
`e.t.trackPageView`, and not in a predicatable way. Because the `trackPageView`
method could be used by indeterminable other scripts the compiler can't minify
it, so testing for just the presence of its name is safe.