Make G5 additional filenames nicer
Some G5 services have additional documents associated with them.

Previously we extracted the supplied filenames from document URLs to use as link text. This worked OK for most documents (when the URLs still had spaces in them), but since fixing the document URLs in the export script these look very bad, with lots of '%20's in there.

Rather than url-decode the URLs to recover the spaces it is better to use a standard name for additional documents, as some filenames are very long and look bad on the page, even with spaces in them instead of '%20's.