Use custom access log for CloudWatch
Create a new Apache access log with type and application information in
so that we can distinguish where a log message has come from. The
CloudWatch metric filters are applied to all messages in a log group and
seeing as we put all our log streams in the same group so that we can
easily search over them this means our alarms do not work properly.

This change also updates the alarm descriptions to be a bit more
helpful. They now say which application and environment the alarm came
from and link to the manual explaining what should be done.