Use content loader
Previously the service presenter used a hard-coded mapping to know which service
attributes to put on the page.

This commit changes the service presenter to use the content loader in dmutils
instead.

It adds a separate manifest for service pages, which is derived from the
previous hard-coded mapping (so should have no effect on which attributes
are displayed)