Support searching for '.' versions of the G5 service IDs
[#99026308](https://www.pivotaltracker.com/story/show/99026308)

When forming a request for the search-api buyer frontend will look
for G5 service IDs in the query string and try to replace dots with
'-'. This way, search-api always gets a request containing '-' IDs.

This allows us to support searches for '.' service IDs without making
any changes to the search-api index.

Only service IDs with all-'.' are suppoted, mixing '.' and '-' will
not trigger the replacement and probably won't return any results.