Patch GOVUK.Analytics to add 'allowLinker' option
The GOVUK.Analytics interface doesn't set the
'allowLinker' option when creating a ga tracker.

In order to preserve a vendor-agnostic
interface it also doesn't allow this to be sent
through when creating a new GOVUK.Analytics
instance.

This commit patches
GOVUK.GoogleAnalyticsUniversalTracker to include
'autoLinker' as a default option for all ga
trackers.