Split `views.py` into a folder of multiple view files
Having all routes in a single `views.py` file is becoming unmanageable as more pages are added to the app.  This new structure mirrors the structure already used in the supplier app.

This alos removes two previously-deprecated redirect routes for legacy URLs for service pages and search from the Grails app. (Deprecation date was flagged as 1st Jan 2016, which we are now beyond).