Pass optional error message to 503 template
The second parameter on the `abort()` call was being tested in one
of our tests, so I figured it was worth preserving the error messages
if possible.

`abort()` accepts three parameters as input:
- `mapping` (ie, status_code)
- `description`
- `response`

If a description is passed in, it overwrites a description provided by default.
([This is the default description for a 503 response](https://github.com/mitsuhiko/werkzeug/blob/master/werkzeug/exceptions.py#L582)).
This means we can't just check *if* a description exists, because it should.

There is no default response, however, so we can just check if one exists and, if
so, pass it into our template.

That's the thinking, anyway.