Remove custom version of gulp-jasmine-phantom
We made a custom version to deal with the fact
that the versio of Node that Travis uses in the VM
image for Python projects was below that required
for the NPM module.

This is no longer the case (it's currently running
at 4.2 (using nvm) so the custom version is no
longer needed.