Add a catalogue of briefs page
This adds a page where all briefs belonging to a framework can be viewed.  It re-uses the styles from the G-Cloud search results page as much as possible.

Both `live` and `closed` briefs are listed, in reverse order of publication date (i.e. with the freshest briefs at the top of the list).