Show closed briefs on buyer dashboard
Previously briefs just disappeared from the buyer dashboard when their "closing time" came around.

This creates a new section on the dashboard that lists closed briefs, along with the number of responses that were received.

It also adds a column for published briefs showing the closing date, as this will be useful information for buyers to see.