Add brief clarification question numbers
[#115944077](https://www.pivotaltracker.com/story/show/115944077)

Numbers are needed in the buyer and public view of the clarification
questions so that later questions can easily refer to earlier questions.

The fact that summary tables are done with macros makes it more
difficult to just use Jinja2's `loop.counter` variable. The simplest way
of getting a counter actually seems to be to pass it in to the template.
While this number could be generated in the API it is just for display,
ideally it would be generated in the view but that is not easy because
of how we use macros.