'Brief summary' -> 'Brief section'
We have all these words floating around and they make it hard
to talk about things in a consistent way.
The grandparent page is no longer a brief summary, but an overview
of the things needed to work through a brief.
The parent page *is* still a summary page, so keep that.