404 if lot_slug in url doesn't match lot slug in brief
There was a pretty silly bug whereby changing the lot slug in the
url of most of the brief pages would be allowed because at no point
were we checking that the `framework_slug` or `lot_slug` in the URL
matched those in the brief.

Added a bunch of new tests and a new helper method that checks the
framework and lot match those in the brief.
Also fixed a few existing tests that were working even though they
shouldn't have been.

Also stopped passing `framework` and `lot` to our brief templates
because none of them needed anything other than what was already
in the `brief_data`.