Set G-Cloud lot filters to size 16
Same thing as was raised in the frontend toolkit a few days past.

At some indefinite period in the history of our apps,
our option select labels were embiggened to
godzilla-like proportions.

While the value text for our option labels has since been reset,
the lot filters themselves are still 19px font.

It seems like all the rest of our filter values are 16px and so
these ones should be as well to maintain the association.

[Relevant SHA in frontend-toolkit](207dc7e196079125ad47a5edf2e4dcbb860077ba)