Adjust button width for new padding
The horizontal padding for all buttons using the
`button` mixin was increased from 1.2em
(0.6em x 2) to 1.578946em (0.789473em x 2) in
version 14.12.0 of the govuk frontend toolkit:

https://github.com/alphagov/govuk_frontend_toolkit/commit/227093d05ea6faa1c3a6dd1aed919ce607b66182

The 'Show services' button has its width set to
7.6em (roughly 183px with the 24px font-size) so
when the padding was increased, this reduced the
space available for the text.

In the CSS for this button, bearing in mind 1px
is 0.04166em (1 / 24), the following is true:

- before the change the text had 6.4em
  (7.6em - 1.2em) of space
- after the change the text had 6.02105em
  (7.6em - 1.578946em)

This changes the width so it includes the space
the text had before + the new padding:

6.4em + 1.57894em = 7.97894em (which we round up
to 8em).