Remove unneeded home class
I have done extensive research on this.

While working on a story to update the header links,
@TheDoubleK asked what the "home" class was for. [1]

The "home" class for header links was introduced quite early
in the history of our apps. [2]
More recent commits have some header links with the "home"
class and some without. [3]

The govuk_prototype_kit doesn't use them. [4]
I thought it might have come from here, but no.
Nor is it mentioned in the official documentation. [5]

Finally, there's no rules for `.home` in our frontend toolkit. [6]
Nor is there in this app. [7]

So let's kill them. @gemmaleigh agrees with me.

[1](https://github.com/alphagov/digitalmarketplace-supplier-frontend/pull/494#discussion_r67329199)
[2](https://github.com/alphagov/digitalmarketplace-supplier-frontend/blob/7f7420ef022405f37aff712f913ba50ca7fbd62c/app/templates/_proposition_header.html)
[3](https://github.com/alphagov/digitalmarketplace-buyer-frontend/commit/756bea36a810476c0b5c5736229a9d10148cb206)
[4](https://github.com/alphagov/govuk_prototype_kit/blob/master/app/views/includes/propositional_navigation.html)
[5](https://github.com/alphagov/govuk_template#propositional-title-and-navigation)
[6](https://github.com/alphagov/digitalmarketplace-frontend-toolkit/search?utf8=%E2%9C%93&q=.home)
[7](https://github.com/alphagov/digitalmarketplace-buyer-frontend/search?utf8=%E2%9C%93&q=.home)