Add /suppliers/<code> endpoint for supplier details page
Loads the supplier from the API and renders the supplier details
page. Returns 404 if supplier with the given code doesn't exist.

(This is using the dmapiclient version of the get_supplier method,
since it's identical to the `app/api_client` right now and the
vendored client is not yet wired up.)