Only remove length from normal flow if GET request
The requirements length question is a required question but needs to be
handled differently on the publish confirmation page. As such, if it's
unanswered and the page is being rendered it needs to be removed from
the count of unanswered questions.

If a POST request is sent to the same route however, it does need to be
included in the count so that the request can be aborted if the question
is unanswered. Moving the check to after the POST check resolves this.