Add condition that brief is not withdrawn to `is_brief_correct`
With a new status being introduced that allows a brief's status to
be "withdrawn", we wish to make sure that withdrawn briefs are
no longer accessible to users. Most pages at the moment would not
show the user a withdrawn brief and would instead 404, however
there are a couple that would. Instead of adding a specific check
to each view, we will add the check to the widely used
`is_brief_correct` function. This should remove the responsibility
that a developer must remember to check if a brief is withdrawn or
not.

Note, `is_brief_correct` is used for when users related to the
brief are interacting with the brief i.e. viewing their brief's
responses. This funciton is not used for parts of the app such as
the marketplace showing all briefs but these parts of the app
already have sufficient brief status checks.

`is_brief_correct` was lacking any unit tests so have also added
these.