Set custom dimension for category filters
For each filter group, we wish to track which filters have been used.
For example: lot = digital outcomes and status = open

We make sure the custom dimension sends back the value of the dimension
in the same order everytime (alphabetical) so google analytics will
be consistent.

We added `search` to the location wrapper which collects params from
the query string which can then be used for our custom dimensions.