Bugfix: show all service ID digits for 15-digit IDs
Previous G-Cloud iterations had 16-digit service IDs, which are displayed on the service page as four chunks of four digits.
For G-Cloud 8 we have moved to 15-digit IDs to help people using spreadsheets, but the method used to "chunk" the IDs for display on the service page dropped the final three digits as they did not match the regular expression that was being used to split the ID (`r'....'`).

This introduces a new method `chunk_string` that returns a generator for chunks of a specified length.  Unlike the old regular expression, this generator also returns smaller chunks at the end of the string, so that 15-digit IDs display as three chunks of four plus three digits at the end (e.g. 1111 2222 3333 444).