Fix bug on publish page with required questions
When the update to the buyer frontend to introduce one or two week
specialist briefs the question was referring to briefs as being 'live'
and not 'open'. This was changed after the app was released which broke
the code which removed that particular question from the count of
unanswered questions. This update is needed to fix it.