don't distinguish between expired password tokens vs invalid tokens
dmutils' new encryption method doesn't distinguish between tokens
older than a day and tokens that it can't decrypt at all - so any
errors that it raises we should consider as expired. A user shouldn't
need to distinguish between the two states - as the correct action in
either case is probably to get a new token issued.