Assert that important dates are equal, not true
The previous test for checking the dates on the opportunity page was
asserting that results of the xpaths and the values they were supposed
to be were both true. They were both strings so they would always be
true. This meant the tests would always pass, regardless of what string
you supplied as an argument.

This updates to check the result of the xpath is equal to the string
supplied.