Split helpers and move brief start pages to buyers/
Brief start pages for each lot have a `/buyers` prefix, so in order
to allow routing between different apps they'd have to be part of
the briefs app.

Unlike every other route in buyers they don't require a login so
we create an additional blueprint (`dos` to match the view name,
but this can be changed later) in `app.buyers` that doesn't require
authorization.

Split the rest of the helpers and associated tests.