Move /buyers URL prefix to the blueprints
This makes sure that all brief URLs start with /buyers.

The only remaining views that use the same prefix outside the buyers
blueprint are `/buyers/create` GET and POST views. Moving them from
main requires duplicating a lot of the forms and helper functions,
and it feels like the right solution for them is a separate users
app.

The simplest thing we can do right now would be to change URLs to
eg `/create/buyer`. But I don't want to introduce any functional
changes with this split, so I'm leaving this for now.