Use a list when passing framework state to helper
You can pass just a string and the `get_framework_and_lot` helper will
work (as long as it's just one allowed status). The problem is if there
is ever a status which is a subset of another you'll get an unexpected
results. For example if there were two statuses `closed` and
`not closed`, `closed` is contained in `not_closed`