Use brief publish date to render correct content
When there are no responses to a brief the content displayed needs to be
different depending on if the brief is a legacy brief (no evidence
required for responses), or a new flow brief (evidence required in
responses).

Previously the split between new and old style briefs was being
determined by the `requirementsMet` key of the responses. If there are
no responses however this breaks down and we need another way to
determine what kind of brief we're dealing with.

This can be done by comparing the `publishedAt` date of the brief with
the feature flag set in the config.

This may not be a permanent solution. We may want to remove that flag at
some point in the future. We could possibly hard code the date the new
flow went live in the logic.