404 brief responses page if brief is not closed
Previously we were somewhat supporting this page for live briefs
even though we didn't link to the page. We have removed this
functionality because there were bugs on the page for live briefs
and we had no evidence that it is worth the maintenance cost as
the page is effecitively hidden unless a user 'hacks' the url.