Check that service is a G-Cloud service
Rather than try...except on loading content for services, which would fail for DOS-style services with no "display_service" manifest, it's better to instead check that the service ID requested is a G-Cloud service.  This avoids the possiblity of swallowing genuine errors, e.g. if a necessary manifest was missing.