Remove plain text logs in live
25.2.0 of utils will create json logs by default, unless the
DM_PLAIN_TEXT_LOGS config variable is set to True in the config.

This config variable is set in dev and test so that logs are more
legible/useful.

The other thing 25.2.0 does is log to stdout instead of file by default,
unless the DM_LOG_PATH config variable is set. In the PaaS
environment, we'll need to log to stdout so will overwrite this
variable. Logging to stdout is also more useful in dev and test.