Create macros for the search category filter
As the category filter is a recursive-ish structure, I've tried
to generalise the elements into small macros that call each other
to build the filter so that we're in theory less tied to a
specific data structure that comes back from the API.

In practice, we haven't ever really used our search filter stuff
(ie, the "option select" pattern) outside of this search page, so
maybe being tied to a data structure isn't the end of a world for
a pattern we may never generalise.

Not sure if `aria-current` is the right one to use, but it seemed
like a reasonable fit.
https://ljwatson.github.io/design-patterns/aria-current