Further abstract the category filter macros
The top-level lot list and each child list of subcategories
differed only ever so slightly in the logic needed to build them,
and even then it was mostly about getting the values from different
keys (ie, the on-screen text for a lot is under `lot.name` whereas
for a category, we need to look in `category.label`).

Slightly more complicated macros remove us still further from the
actual data structure coming back from the API.  It's not totally
abstract (for example, category key names are still hardcoded in
the `item_list__selected` macro), but it seems like it's pretty
far along.

Should we need to generalise this further, I would suggest that
it can be done the next time we need one of these lists.