Add visually hidden back information
Psuedo-elements don't show up in the DOM, so screenreaders won't
pick up on the fact that there are arrows pointing backwards.

The arrows pointing backwards are just telling us that we're moving
backwards (or upwards) in the hierarchy, so I've added in some
text that makes this explicit.

There might be a more elegant solution to this.