Add "last" class to deepest category list
We want all of the lists to sit flush along the left side of the
column except the final list, which we want to show as indented.

CSS doesn't have great tools to accomplish this to begin with,
and it doesn't help us at all when we don't know how deeply nested
the lists are going to be.

Came up with a jinja variable that can add a "last" class to a list
by checking that:

- either no items in the current list contain nested items
- or, if they do, that no nested items are "selected"