Refactor CSS
Remove all the inline CSS and rewrite the SCSS that applies to this
list.

Now that the CSS is isolated and the category list is building
itself using macros, we have a pretty isolate pattern we can move
out of this app and into the frontend toolkit.

Note: I've created a SCSS variable that was originally 30px (the
same as our `$gutter`), but I think looks better at 24px.
The reason that I didn't just use the `$gutter` variable and its
derivations (`$gutter-half`, `$gutter-one-third`, etc) is because
they didn't really seem fine-grained enough for this pattern; we
have a lot of items in pretty close contact and only being able to
use multiples of `5px` made the filter look generally worse.