Update brief fixtures
- We update the brief fixture to be on the latest framework.
- We update the format of `startDate` to be saved in the parseable
date format rather than just a simple string.
- We add keys that usually exist on a brief that were not included,
such as `technicalWeighting` and `requirementsLength`.
- We add keys related to the clarification question time period.
These before had not been added and it was therefore assumed
that the brief was open for clarification questions. To make this
more explicit we give dates for `clarificationQuestionsClosedAt`
and `clarificationQuestionsPublishedBy` that are realistic compared
to the `publishedAt` date.

This means that our fixture now represents a brief it were to
be published 'today' and our tests should have more realistic
coverage.

For the multiple brief fixtures:
- Order keys alphabetically for better readability
- Add keys that would normally exist on a brief but were missing
from our stubs
- Update `startDate` to use our new dateformat for one of the
briefs