Remove views for creating buyer users
The `create_buyer_user` and `submit_create_buyer_user` views have been
moved to the new briefs app. They both used the URL `/buyers/create`
which clashed with the routing for the new briefs app. Any requests for
that URL were directed to briefs app due to the `/buyers` prefix.

Rather than changing the URL the views have been moved. As they don't
acutally create users, just send an email, it's ok to move them with the
briefs app.

User creation is the roadmap to be split into a separate app, so all
these views will be looked at again in the near future.