Create `external` blueprint for external links
The buyer app needs to be able to link to the views moved to the new
briefs app.

Rather than hardcoding the links in, by creating the external blueprint
we can add a view to match any external views we need and then still use
`url_for` to generate URL's for the templates.