Preserve line breaks for textarea formatted text
We're going to allow newline formatting for `textbox_large`
content submitted to us and displayed back to users.

In this pull request, we're allowing newlines in:

- textbox_large areas of individual brief pages
  (brief summary, workplace description, etc.)
- clarification questions and answers for briefs
- textbox_large areas of individual G8/G9 service pages
  (includes most of the follow-up questions)

Where we can, we're checking if the question type is
`textbox_large` before applying the filter. But on the
brief questions and answers, for example, we don't have a
question object to inspect -- only a list of strings. We know
they're all textbox large questions anyway, so there we just
use the filter without checking.

We are *not* adding this to search results because search
result rankings are a competitive space and we don't want there
to be an incentive to create very large service summaries and/or
brief descriptions.