Changes wording on Keyword Search from "Keywords" to "Keyword Search".
Ticket: https://trello.com/b/rdgM0WNN/welcome-to-gds-step-1-make-a-copy-of-this-board-and-use-that

Adds in "Apply Filters" header for filters.

    Ticket: https://trello.com/b/rdgM0WNN/welcome-to-gds-step-1-make-a-copy-of-this-board-and-use-that

Changing "Keyword Search" to "Keyword search"

Changes "Apply Filters" to "Apply filters"

Ticket: https://trello.com/c/P7HpqQAU/552-add-a-heading-above-the-filters-and-make-the-headings-consistent

Adds a 5px margin to the bottom of the "Apply filters" title to be more consistent with other titles.

Ticket: https://trello.com/c/P7HpqQAU/552-add-a-heading-above-the-filters-and-make-the-headings-consistent

changes margin to 10px