Implement and satisfy flake8
Flake 8 is really useful for finding unused imports which there could be
a lot of after stripping out an app.

The max complexity has been set to 12. This was the highest complexity
found at the time of implementation. This way, it should stop any new
code from going higher whilst maintaining test functionality.