Convert `list opportunites` to use elasticsearch
The view for listing briefs was previously using postgres. This updates
it to use elasticsearch instead. It uses the same pattern and shares a
lot of code with the g-cloud search view.

As we do not show users briefs with the `withdrawn` status, if a q
status filter is not selected, we update the request to search by all
status _except_ `withdrawn`. This prevents it ever being retrieved.