Correctly build category tree if lot has no categories
With g-cloud, once a lot is selected the others are not included in the
category tree. This is because g-cloud lots have categories to display,
and having those as well as the other lots would be visually confusing.

With DOS, we have no categories with a lot, and so it make sense to
include all lots in the root node, even after one has been selected.