Use and refactor search templates for opportunities
Use the existing search templates for g-cloud and apply them to DOS
opportunities listing.

A lot of the templates can be shared and so a lot of prefixes have been
dropped.

The `_briefs_results` template has been modified a reasonable amount to
work with the new data structure for briefs. This also includes the use
of a new jinja filter, to convert snake cased specialist role names into
normal text.